features/eu-debug: Initial patches for eu-debug support for LNL/BMG
eu-debugger preview patches on top of 6.11-rc6. eu-debug merge of drm-tip
is going to take some time so to avoid frequent uapi break, prelim uapi
is being used and it will be changed during merge of drm-tip.

Patches are under review in drm mailinglist at
https://patchwork.freedesktop.org/series/136572/ and is being updated as
review continue, we are maintaining feature at
https://gitlab.freedesktop.org/miku/kernel/-/tree/eudebug-dev

Signed-off-by: Kolanupaka Naveena <kolanupaka.naveena@intel.com>