Let patches relative to series file
Don't assume the kernel location is in a inner directory,
rather always let the patches relative to the series file
so the caller can adjust. This makes it compatible with
quilt and git-pile. Example, with a pre-existent checkout:

	$ git pile --no-config genbranch --no-cache --inplace \
		-e ../xekmd-backports/ \
		-x v6.11-rc6

backport.sh is changed to consider that so it continues to
work.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>