backport/base: Add engine disable support and configfs interface
Introduce support to selectively disable engines for targeted debugging and
early enabling scenarios. Engines can now be disabled in software, in addition
to hardware fuse checks. A new configfs interface is added to allow userspace
to control which engines are enabled:

    echo 'rcs0,bcs0' > /sys/kernel/config/xe/<bdf>/engine_allowed

This enables only the specified engines, while others remain disabled. A '*'
can be used to enable all instances of a class.

Additionally:
 - Converted "fused off" messages to be GT-based for better multi-tile logging.
 - Added internal configfs API documentation and corrected header formatting.
 - Dropped unnecessary semicolons from dummy configfs implementations.

Signed-off-by: Bommu Krishnaiah <krishnaiah.bommu@intel.com>